Extract server injection functions
